modularize renderTimeLimit into stage.utils shared by survey/privateChat/groupChat
