+++ docs/assets/api/schemas.json
@@ -2539,14 +2539,6 @@
         "anonymousProfileSetId": {
           "type": "string"
         },
-        "timeLimitInMinutes": {
-          "minimum": 1,
-          "type": ["integer", "null"]
-        },
-        "timeMinimumInMinutes": {
-          "minimum": 1,
-          "type": ["integer", "null"]
-        },
         "questions": {
           "type": "array",
           "items": {
@@ -2565,6 +2557,14 @@
               }
             ]
           }
+        },
+        "timeLimitInMinutes": {
+          "minimum": 1,
+          "type": ["integer", "null"]
+        },
+        "timeMinimumInMinutes": {
+          "minimum": 1,
+          "type": ["integer", "null"]
         }
       }
     },
