+++ frontend/src/components/stages/group_chat_editor.ts
@@ -6,8 +6,8 @@ import '@material/web/checkbox/checkbox.js';
 
 import {core} from '../../core/core';
 import {AuthService} from '../../services/auth.service';
-
 import {ExperimentEditor} from '../../services/experiment.editor';
+import {renderTimeLimit} from '../../shared/stage.utils';
 
 import {ChatStageConfig} from '@deliberation-lab/utils';
 
@@ -30,8 +30,12 @@ export class ChatEditor extends MobxLitElement {
 
     return html`
       <div class="title">Conversation settings</div>
-      ${this.renderTimeLimit()} ${this.renderTurnBasedSetting()}
-      ${this.renderReactionsSetting()}
+      ${renderTimeLimit({
+        stage: this.stage,
+        canEdit: this.experimentEditor.canEditStages,
+        onStageChange: (stage) => this.experimentEditor.updateStage(stage),
+      })}
+      ${this.renderTurnBasedSetting()} ${this.renderReactionsSetting()}
       <div class="divider"></div>
       <div class="title">Agent mediators</div>
       <div class="description">
@@ -118,94 +122,6 @@ export class ChatEditor extends MobxLitElement {
             </div>
           `,
         )}
-      </div>
-    `;
-  }
-
-  private renderTimeLimit() {
-    const timeLimit = this.stage?.timeLimitInMinutes;
-
-    const updateCheck = () => {
-      const isSet = this.stage?.timeLimitInMinutes != null;
-      this.experimentEditor.updateStage({
-        ...this.stage!,
-        timeLimitInMinutes: isSet ? null : 20,
-        timeMinimumInMinutes: null,
-      });
-    };
-
-    const updateMaxTime = (e: InputEvent) => {
-      const val = (e.target as HTMLInputElement).valueAsNumber;
-      const timeLimitInMinutes = val > 0 ? Math.floor(val) : null;
-      this.experimentEditor.updateStage({
-        ...this.stage!,
-        timeLimitInMinutes,
-      });
-    };
-
-    const updateMinTime = (e: InputEvent) => {
-      const val = (e.target as HTMLInputElement).valueAsNumber;
-      const minTime = val > 0 ? Math.floor(val) : null;
-      const max = this.stage?.timeLimitInMinutes;
-      const timeMinimumInMinutes =
-        minTime != null && max != null ? Math.min(minTime, max) : minTime;
-      this.experimentEditor.updateStage({
-        ...this.stage!,
-        timeMinimumInMinutes,
-      });
-    };
-
-    return html`
-      <div class="config-item">
-        <div class="checkbox-wrapper">
-          <md-checkbox
-            touch-target="wrapper"
-            ?checked=${timeLimit != null}
-            ?disabled=${!this.experimentEditor.canEditStages}
-            @click=${updateCheck}
-          >
-          </md-checkbox>
-          <div>Disable conversation after a fixed amount of time</div>
-        </div>
-        ${timeLimit != null
-          ? html`
-              <div class="number-input tab">
-                <label for="timeLimit">
-                  Maximum time in minutes (starting at first message).
-                  Participant will remain in chat until minimum messages
-                  requirement is met, even if maximum time has passed.
-                </label>
-                <input
-                  type="number"
-                  id="timeLimit"
-                  name="timeLimit"
-                  min="1"
-                  step="1"
-                  .value=${timeLimit}
-                  ?disabled=${!this.experimentEditor.canEditStages}
-                  @input=${updateMaxTime}
-                />
-              </div>
-              <div class="number-input tab tab-bottom">
-                <label for="timeMinimum">
-                  Minimum time participants must stay (in minutes). Takes
-                  precedence over maximum number of messages.
-                </label>
-                <input
-                  type="number"
-                  id="timeMinimum"
-                  name="timeMinimum"
-                  min="1"
-                  step="1"
-                  .max=${timeLimit ?? ''}
-                  .value=${this.stage?.timeMinimumInMinutes ?? ''}
-                  placeholder="No minimum"
-                  ?disabled=${!this.experimentEditor.canEditStages}
-                  @input=${updateMinTime}
-                />
-              </div>
-            `
-          : nothing}
       </div>
     `;
   }
