+++ frontend/src/shared/stage.utils.ts
@@ -1,3 +1,4 @@
+import {html, nothing} from 'lit';
 import {ChatMessage, UnifiedTimestamp} from '@deliberation-lab/utils';
 import {Timestamp} from 'firebase/firestore';
 
@@ -38,4 +39,115 @@ export function getChatTimeRemainingInSeconds(
   const elapsed = Timestamp.now().seconds - startTimestamp.seconds;
   const remaining = chatStage.timeLimitInMinutes * 60 - elapsed;
   return remaining > 0 ? Math.floor(remaining) : 0;
+}
+
+export interface StageWithTimeLimit {
+  timeLimitInMinutes: number | null;
+  timeMinimumInMinutes: number | null;
+}
+
+export interface RenderTimeLimitOptions<T extends StageWithTimeLimit> {
+  stage: T | undefined;
+  canEdit: boolean;
+  onStageChange: (stage: T) => void;
+  checkboxTitle?: string;
+  maxTimeLabel?: string;
+  minTimeLabel?: string;
+}
+
+/**
+ * Renders the stage time limit editor controls (toggle, max time, min time).
+ */
+export function renderTimeLimit<T extends StageWithTimeLimit>(
+  options: RenderTimeLimitOptions<T>,
+) {
+  const {
+    stage,
+    canEdit,
+    onStageChange,
+    checkboxTitle = 'Disable conversation after a fixed amount of time',
+    maxTimeLabel = 'Maximum time in minutes (starting at first message). Participant will remain in chat until minimum messages requirement is met, even if maximum time has passed.',
+    minTimeLabel = 'Minimum time participants must stay (in minutes). Takes precedence over maximum number of messages.',
+  } = options;
+
+  if (!stage) return nothing;
+
+  const timeLimit = stage.timeLimitInMinutes;
+
+  const updateCheck = () => {
+    const isSet = stage.timeLimitInMinutes != null;
+    onStageChange({
+      ...stage,
+      timeLimitInMinutes: isSet ? null : 20,
+      timeMinimumInMinutes: null,
+    });
+  };
+
+  const updateMaxTime = (e: InputEvent) => {
+    const val = (e.target as HTMLInputElement).valueAsNumber;
+    const timeLimitInMinutes = val > 0 ? Math.floor(val) : null;
+    onStageChange({
+      ...stage,
+      timeLimitInMinutes,
+    });
+  };
+
+  const updateMinTime = (e: InputEvent) => {
+    const val = (e.target as HTMLInputElement).valueAsNumber;
+    const minTime = val > 0 ? Math.floor(val) : null;
+    const max = stage.timeLimitInMinutes;
+    const timeMinimumInMinutes =
+      minTime != null && max != null ? Math.min(minTime, max) : minTime;
+    onStageChange({
+      ...stage,
+      timeMinimumInMinutes,
+    });
+  };
+
+  return html`
+    <div class="config-item">
+      <div class="checkbox-wrapper">
+        <md-checkbox
+          touch-target="wrapper"
+          ?checked=${timeLimit != null}
+          ?disabled=${!canEdit}
+          @click=${updateCheck}
+        >
+        </md-checkbox>
+        <div>${checkboxTitle}</div>
+      </div>
+      ${timeLimit != null
+        ? html`
+            <div class="number-input tab">
+              <label for="timeLimit">${maxTimeLabel}</label>
+              <input
+                type="number"
+                id="timeLimit"
+                name="timeLimit"
+                min="1"
+                step="1"
+                .value=${timeLimit}
+                ?disabled=${!canEdit}
+                @input=${updateMaxTime}
+              />
+            </div>
+            <div class="number-input tab tab-bottom">
+              <label for="timeMinimum">${minTimeLabel}</label>
+              <input
+                type="number"
+                id="timeMinimum"
+                name="timeMinimum"
+                min="1"
+                step="1"
+                .max=${timeLimit ?? ''}
+                .value=${stage.timeMinimumInMinutes ?? ''}
+                placeholder="No minimum"
+                ?disabled=${!canEdit}
+                @input=${updateMinTime}
+              />
+            </div>
+          `
+        : nothing}
+    </div>
+  `;
 }
